moved the setup file back to its original place
